refactor: modularize application
