+++ app.py
@@ -74,12 +74,6 @@ N_WORKERS_DEFAULT  = max(1, min((os.cpu_count() or 2) - 1, 6))
 MARGIN_FAC_DEFAULT = 1.15
 OVERWRITE_DEFAULT  = False
 
-LAYER_AE_NAME = "Aire d'étude élargie"
-LAYER_ZE_NAME = "Zone d'étude"
-
-BASE_SHARE = r"\\192.168.1.240\commun\PARTAGE"
-SUBPATH    = r"Espace_RWO\CARTO ROBIN"
-
 OUT_IMG    = r"C:\Users\utilisateur\Mon Drive\1 - Bota & Travail\+++++++++  BOTA  +++++++++\---------------------- 3) BDD\PYTHON\2) Contexte éco\OUTPUT"
 
 # Dossier par défaut pour la sélection des shapefiles (onglet 1)
@@ -117,54 +111,8 @@ COMMENT_TEMPLATE = (
 )
 
 # Onglet 3 — Identification Pl@ntNet
-API_KEY = "2b10vfT6MvFC2lcAzqG1ZMKO"  # Votre clé API Pl@ntNet
-PROJECT = "all"
-API_URL = f"https://my-api.plantnet.org/v2/identify/{PROJECT}?api-key={API_KEY}"
 
-# =========================
-# Utils communs
-# =========================
-def log_with_time(msg: str) -> None:
-    print(f"[{datetime.datetime.now().strftime('%H:%M:%S')}] {msg}")
-
-def normalize_name(s: str) -> str:
-    s2 = s.replace("\u00A0", " ").replace("\u202F", " ")
-    while "  " in s2:
-        s2 = s2.replace("  ", " ")
-    return s2.strip().lower()
-
-def to_long_unc(path: str) -> str:
-    if path.startswith("\\\\?\\"): return path
-    if path.startswith("\\\\"):   return "\\\\?\\UNC" + path[1:]
-    return "\\\\?\\" + path
-
-def chunk_even(lst: List[str], k: int) -> List[List[str]]:
-    if not lst: return []
-    k = max(1, min(k, len(lst)))
-    base = len(lst) // k
-    extra = len(lst) % k
-    out, start = [], 0
-    for i in range(k):
-        size = base + (1 if i < extra else 0)
-        out.append(lst[start:start+size])
-        start += size
-    return out
-
-def load_prefs() -> dict:
-    if os.path.isfile(PREFS_PATH):
-        try:
-            with open(PREFS_PATH, "r", encoding="utf-8") as f:
-                return json.load(f)
-        except Exception:
-            return {}
-    return {}
-
-def save_prefs(d: dict) -> None:
-    try:
-        with open(PREFS_PATH, "w", encoding="utf-8") as f:
-            json.dump(d, f, ensure_ascii=False, indent=2)
-    except Exception:
-        pass
+from modules.utils import (log_with_time, normalize_name, to_long_unc, chunk_even, load_prefs, save_prefs)
 
 class TextRedirector:
     def __init__(self, widget): self.widget = widget
@@ -194,334 +142,11 @@ class ToolTip:
         self._cancel()
         if self.tipwindow: self.tipwindow.destroy(); self.tipwindow = None
 
-# =========================
-# Fonctions Pl@ntNet
-# =========================
-def resize_image(image_path, max_size=(800, 800), quality=70):
-    """
-    Redimensionne et compresse une image.
+from modules.plantnet import resize_image, identify_plant, copy_and_rename_file
 
-    :param image_path: Chemin de l'image à traiter.
-    :param max_size: Tuple indiquant la taille maximale (largeur, hauteur).
-    :param quality: Qualité de compression (1-100).
-    :return: BytesIO de l'image traitée ou None en cas d'erreur.
-    """
-    try:
-        with Image.open(image_path) as img:
-            img.thumbnail(max_size)
-            buffer = BytesIO()
-            img.save(buffer, format='JPEG', quality=quality)
-            buffer.seek(0)
-            return buffer
-    except Exception as e:
-        print(f"Erreur lors du redimensionnement de l'image : {e}")
-        return None
+from modules.context_eco import worker_run, discover_projects, LAYER_AE_NAME, LAYER_ZE_NAME, BASE_SHARE, SUBPATH
 
-def identify_plant(image_path, organ):
-    """
-    Envoie une image à l'API Pl@ntNet pour identification.
-
-    :param image_path: Chemin de l'image à envoyer.
-    :param organ: Type d'organe de la plante (par exemple, 'flower').
-    :return: Nom scientifique de la plante identifiée ou None.
-    """
-    print(f"Envoi de l'image à l'API : {image_path}")
-    try:
-        resized_image = resize_image(image_path)
-        if not resized_image:
-            print(f"Échec du redimensionnement de l'image : {image_path}")
-            return None
-
-        files = {
-            'images': (os.path.basename(image_path), resized_image, 'image/jpeg')
-        }
-        data = {
-            'organs': organ
-        }
-
-        response = requests.post(API_URL, files=files, data=data)
-
-        print(f"Réponse de l'API : {response.status_code}")
-        if response.status_code == 200:
-            json_result = response.json()
-            try:
-                species = json_result['results'][0]['species']['scientificNameWithoutAuthor']
-                print(f"Plante identifiée : {species}")
-                return species
-            except (KeyError, IndexError):
-                print(f"Aucun résultat trouvé pour l'image : {image_path}")
-                return None
-        else:
-            print(f"Erreur API : {response.status_code} - {response.text}")
-            return None
-    except Exception as e:
-        print(f"Exception lors de l'identification de la plante : {e}")
-        return None
-
-def copy_and_rename_file(file_path, dest_folder, new_name, count):
-    """
-    Copie et renomme un fichier dans le dossier de destination.
-
-    :param file_path: Chemin du fichier original.
-    :param dest_folder: Dossier de destination.
-    :param new_name: Nom scientifique de la plante.
-    :param count: Compteur pour différencier les fichiers portant le même nom.
-    """
-    ext = os.path.splitext(file_path)[1]
-    if count == 1:
-        new_file_name = f"{new_name} @plantnet{ext}"
-    else:
-        new_file_name = f"{new_name} @plantnet({count}){ext}"
-    new_path = os.path.join(dest_folder, new_file_name)
-    try:
-        shutil.copy(file_path, new_path)
-        print(f"Fichier copié et renommé : {file_path} -> {new_path}")
-    except Exception as e:
-        print(f"Erreur lors de la copie du fichier : {e}")
-
-# =========================
-# Fonctions QGIS (onglet 1) — inchangées
-# =========================
-def worker_run(args: Tuple[List[str], dict]) -> Tuple[int, int]:
-    projects, cfg = args
-
-    os.environ["OSGEO4W_ROOT"] = cfg["QGIS_ROOT"]
-    os.environ["QGIS_PREFIX_PATH"] = cfg["QGIS_APP"]
-    os.environ.setdefault("GDAL_DATA", os.path.join(cfg["QGIS_ROOT"], "share", "gdal"))
-    os.environ.setdefault("PROJ_LIB", os.path.join(cfg["QGIS_ROOT"], "share", "proj"))
-    os.environ.setdefault("QT_QPA_FONTDIR", r"C:\Windows\Fonts")
-
-    qt_base = None
-    for name in ("Qt6", "Qt5"):
-        base = os.path.join(cfg["QGIS_ROOT"], "apps", name)
-        if os.path.isdir(base):
-            qt_base = base
-            break
-    if qt_base is None:
-        raise RuntimeError("Qt introuvable")
-
-    platform_dir = os.path.join(qt_base, "plugins", "platforms")
-    os.environ["QT_PLUGIN_PATH"] = os.path.join(qt_base, "plugins")
-    os.environ["QT_QPA_PLATFORM_PLUGIN_PATH"] = platform_dir
-    qpa = "windows" if os.path.isfile(os.path.join(platform_dir, "qwindows.dll")) \
-        else ("minimal" if os.path.isfile(os.path.join(platform_dir, "qminimal.dll")) else "offscreen")
-    os.environ["QT_QPA_PLATFORM"] = qpa
-
-    os.environ["PATH"] = os.pathsep.join([
-        os.path.join(qt_base, "bin"),
-        os.path.join(cfg["QGIS_APP"], "bin"),
-        os.path.join(cfg["QGIS_ROOT"], "bin"),
-        os.environ.get("PATH", ""),
-    ])
-
-    sys.path.insert(0, os.path.join(cfg["QGIS_APP"], "python"))
-    sys.path.insert(0, os.path.join(cfg["QGIS_ROOT"], "apps", cfg["PY_VER"], "Lib", "site-packages"))
-
-    from qgis.core import (
-        QgsApplication, QgsProject, QgsLayoutExporter, QgsLayoutItemMap, QgsRectangle,
-        QgsCoordinateTransform
-    )
-
-    qgs = QgsApplication([], False)
-    qgs.setPrefixPath(cfg["QGIS_APP"], True)
-    qgs.initQgis()
-
-    ok = 0
-    ko = 0
-
-    def adjust_extent_to_item_ratio(ext: 'QgsRectangle', target_ratio: float, margin: float) -> 'QgsRectangle':
-        if ext.width() <= 0 or ext.height() <= 0:
-            return ext
-        cx, cy = ext.center().x(), ext.center().y()
-        w, h = ext.width(), ext.height()
-        if (w / h) > target_ratio:
-            new_h = w / target_ratio
-            dh = (new_h - h) / 2.0
-            xmin, xmax = ext.xMinimum(), ext.xMaximum()
-            ymin, ymax = cy - h/2.0 - dh, cy + h/2.0 + dh
-        else:
-            new_w = h * target_ratio
-            dw = (new_w - w) / 2.0
-            ymin, ymax = ext.yMinimum(), ext.yMaximum()
-            xmin, xmax = cx - w/2.0 - dw, cx + w/2.0 + dw
-        cw, ch = (xmax - xmin), (ymax - ymin)
-        mx, my = (margin - 1.0) * cw / 2.0, (margin - 1.0) * ch / 2.0
-        return QgsRectangle(xmin - mx, ymin - my, xmax + mx, ymax + my)
-
-    def extent_in_project_crs(prj: 'QgsProject', lyr) -> Optional['QgsRectangle']:
-        ext = lyr.extent()
-        try:
-            if lyr.crs() != prj.crs():
-                ct = QgsCoordinateTransform(lyr.crs(), prj.crs(), prj)
-                ext = ct.transformBoundingBox(ext)
-        except Exception:
-            pass
-        return ext
-
-    def apply_extent_and_export(layout, lyr_extent: 'QgsRectangle', out_png: str) -> bool:
-        maps = [it for it in layout.items() if isinstance(it, QgsLayoutItemMap)]
-        if not maps: return False
-        for m in maps:
-            size = m.sizeWithUnits()
-            target_ratio = max(1e-9, float(size.width()) / float(size.height()))
-            adj_extent = adjust_extent_to_item_ratio(lyr_extent, target_ratio, cfg["MARGIN_FAC"])
-            m.setExtent(adj_extent); m.refresh()
-        img = QgsLayoutExporter.ImageExportSettings()
-        img.dpi = cfg["DPI"]
-        for attr in ("antialiasing", "antiAliasing"):
-            if hasattr(img, attr):
-                setattr(img, attr, True)
-        try:
-            flag_val = 0
-            for name in dir(img.__class__):
-                if "UseAdvancedEffects" in name:
-                    flag_val |= int(getattr(img.__class__, name))
-            if flag_val and hasattr(img, "flags"): img.flags = flag_val
-        except Exception:
-            pass
-        if hasattr(img, "generateWorldFile"): img.generateWorldFile = False
-        exp = QgsLayoutExporter(layout)
-        res = exp.exportToImage(out_png, img)
-        return res == QgsLayoutExporter.Success
-
-    def relink_layer(prj: 'QgsProject', layer_name: str, shp_path: str) -> Optional[object]:
-        layers = prj.mapLayersByName(layer_name)
-        if not layers: return None
-        lyr = layers[0]
-        try:
-            lyr.setDataSource(shp_path, lyr.name(), "ogr")
-            return lyr
-        except Exception:
-            return None
-
-    def export_views(projet_path: str) -> Tuple[int, int]:
-        okc = 0; koc = 0
-        nom = os.path.splitext(os.path.basename(projet_path))[0]
-        out_ae = os.path.join(cfg["OUT_IMG"], f"{nom}__AE.png")
-        out_ze = os.path.join(cfg["OUT_IMG"], f"{nom}__ZE.png")
-
-        mode = cfg.get("CADRAGE_MODE", "BOTH")
-        expected_exports = (1 if mode in ("AE", "BOTH") else 0) + (1 if mode in ("ZE", "BOTH") else 0)
-
-        prj = QgsProject.instance(); prj.clear()
-
-        opened = False
-        for pth in (projet_path, to_long_unc(projet_path)):
-            try:
-                if prj.read(pth): opened = True; break
-            except Exception:
-                pass
-        if not opened:
-            return 0, expected_exports
-
-        lm = prj.layoutManager()
-        layouts = lm.layouts()
-        if not layouts:
-            prj.clear(); return 0, expected_exports
-        layout = layouts[0]
-
-        lyr_ae = relink_layer(prj, cfg["LAYER_AE_NAME"], cfg["AE_SHP"])
-        lyr_ze = relink_layer(prj, cfg["LAYER_ZE_NAME"], cfg["ZE_SHP"])
-
-        if mode in ("AE", "BOTH"):
-            if (not cfg["OVERWRITE"]) and os.path.exists(out_ae):
-                okc += 1
-            else:
-                if lyr_ae:
-                    ext_ae = extent_in_project_crs(prj, lyr_ae)
-                    if ext_ae and apply_extent_and_export(layout, ext_ae, out_ae): okc += 1
-                    else: koc += 1
-                else:
-                    koc += 1
-
-        if mode in ("ZE", "BOTH"):
-            if (not cfg["OVERWRITE"]) and os.path.exists(out_ze):
-                okc += 1
-            else:
-                if lyr_ze:
-                    ext_ze = extent_in_project_crs(prj, lyr_ze)
-                    if ext_ze and apply_extent_and_export(layout, ext_ze, out_ze): okc += 1
-                    else: koc += 1
-                else:
-                    koc += 1
-
-        prj.clear()
-        return okc, koc
-
-    for p in projects:
-        ok_c, ko_c = export_views(p)
-        ok += ok_c; ko += ko_c
-
-    qgs.exitQgis()
-    return ok, ko
-
-def discover_projects() -> List[str]:
-    partage = BASE_SHARE
-    base_dir: Optional[str] = None
-    try:
-        if os.path.isdir(partage):
-            for e in os.listdir(partage):
-                if normalize_name(e) == normalize_name("Espace_RWO"):
-                    base_espace = os.path.join(partage, e)
-                    for s in os.listdir(base_espace):
-                        if normalize_name(s) == normalize_name("CARTO ROBIN"):
-                            base_dir = os.path.join(base_espace, s); break
-                    break
-    except Exception as e:
-        log_with_time(f"Accès PARTAGE impossible via listdir: {e}")
-
-    if not base_dir or not os.path.isdir(base_dir):
-        base_dir = os.path.join(BASE_SHARE, SUBPATH)
-
-    for d in (base_dir, to_long_unc(base_dir)):
-        try:
-            if not os.path.isdir(d): continue
-            files = os.listdir(d)
-            qgz = [f for f in files if f.lower().endswith(".qgz")]
-            qgz = [f for f in qgz if normalize_name(f).startswith(normalize_name("Contexte éco -"))]
-            return [os.path.join(d, f) for f in sorted(qgz)]
-        except Exception:
-            continue
-    return []
-
-# =========================
-# Fonctions IGN (onglet 2) — identiques au script source
-# =========================
-def dms_to_dd(text: str) -> float:
-    pat = r"(\d{1,3})[°d]\s*(\d{1,2})['m]\s*([\d\.]+)[\"s]?\s*([NSEW])"
-    alt = r"(\d{1,3})\s+(\d{1,2})\s+([\d\.]+)\s*([NSEW])"
-    m = re.search(pat, text, re.I) or re.search(alt, text, re.I)
-    if not m:
-        raise ValueError(f"Format DMS invalide : {text}")
-    deg, mn, sc, hemi = m.groups()
-    dd = float(deg) + float(mn)/60 + float(sc)/3600
-    return -dd if hemi.upper() in ("S", "W") else dd
-
-def add_hyperlink(paragraph, url: str, text: str, italic: bool = True):
-    part = paragraph.part
-    r_id = part.relate_to(
-        url,
-        "http://schemas.openxmlformats.org/officeDocument/2006/relationships/hyperlink",
-        is_external=True
-    )
-    from docx.oxml import OxmlElement, ns
-    fld_simple = OxmlElement('w:hyperlink')
-    fld_simple.set(ns.qn('r:id'), r_id)
-
-    run = OxmlElement('w:r')
-    r_pr = OxmlElement('w:rPr')
-    if italic:
-        i = OxmlElement('w:i')
-        r_pr.append(i)
-    u = OxmlElement('w:u')
-    u.set(ns.qn('w:val'), 'single')
-    r_pr.append(u)
-    run.append(r_pr)
-    run_text = OxmlElement('w:t')
-    run_text.text = text
-    run.append(run_text)
-    fld_simple.append(run)
-    paragraph._p.append(fld_simple)
+from modules.remonter import dms_to_dd, add_hyperlink
 
 # =========================
 # UI — Styles communs
@@ -791,7 +416,7 @@ class ExportCartesTab(ttk.Frame):
             "DPI": int(self.dpi_var.get()),
             "N_WORKERS": int(self.workers_var.get()),
             "MARGIN_FAC": float(self.margin_var.get()),
-        }); save_prefs(self.prefs)
+        }); save_prefs(PREFS_PATH, self.prefs)
 
         t = threading.Thread(target=self._run_export_logic, args=(projets,))
         t.daemon = True; t.start()
@@ -1102,7 +727,7 @@ class RemonterLeTempsTab(ttk.Frame):
                 "RLT_WAIT": float(self.wait_var.get()),
                 "RLT_OUT": self.out_dir_var.get().strip(),
                 "RLT_HEADLESS": bool(self.headless_var.get()),
-            }); save_prefs(self.prefs)
+            }); save_prefs(PREFS_PATH, self.prefs)
 
             print(f"[IGN] Parsing coordonnées…", file=self.stdout_redirect)
             # La chaîne doit contenir LAT puis LON en DMS, séparés par espace/virgule/tab
@@ -1307,7 +932,7 @@ class PlantNetTab(ttk.Frame):
             return
 
         self.prefs["PLANTNET_FOLDER"] = folder
-        save_prefs(self.prefs)
+        save_prefs(PREFS_PATH, self.prefs)
 
         image_extensions = ['.jpg', '.jpeg', '.png', '.heic', '.heif']
         image_files = []
@@ -1653,7 +1278,7 @@ class ContexteEcoTab(ttk.Frame):
             "DPI": int(self.dpi_var.get()),
             "N_WORKERS": int(self.workers_var.get()),
             "MARGIN_FAC": float(self.margin_var.get()),
-        }); save_prefs(self.prefs)
+        }); save_prefs(PREFS_PATH, self.prefs)
 
         t = threading.Thread(target=self._run_export_logic, args=(projets,), daemon=True)
         t.start()
@@ -1723,7 +1348,7 @@ class ContexteEcoTab(ttk.Frame):
             "ZE_SHP": ze,
             "AE_SHP": ae,
             "ID_TAMPON_KM": float(self.buffer_var.get()),
-        }); save_prefs(self.prefs)
+        }); save_prefs(PREFS_PATH, self.prefs)
 
         t = threading.Thread(target=self._run_id_logic, args=(ae, ze, float(self.buffer_var.get())), daemon=True)
         t.start()
@@ -1758,7 +1383,7 @@ class MainApp:
         self.root.title("Contexte éco — Outils")
         self.root.geometry("1060x760"); self.root.minsize(900, 640)
 
-        self.prefs = load_prefs()
+        self.prefs = load_prefs(PREFS_PATH)
         self.style_helper = StyleHelper(root, self.prefs)
         self.theme_var = tk.StringVar(value=self.prefs.get("theme", "light"))
         self.style_helper.apply(self.theme_var.get())
@@ -1797,19 +1422,12 @@ class MainApp:
         new_theme = themes[(themes.index(current) + 1) % len(themes)]
         self.theme_var.set(new_theme)
         self.prefs["theme"] = new_theme
-        save_prefs(self.prefs)
+        save_prefs(PREFS_PATH, self.prefs)
         self.style_helper.apply(new_theme)
 
     def _on_close(self):
         try:
-            save_prefs(self.prefs)
+            save_prefs(PREFS_PATH, self.prefs)
         finally:
             self.root.destroy()
 
-# =========================
-# Main
-# =========================
-if __name__ == "__main__":
-    root = tk.Tk()
-    app = MainApp(root)
-    root.mainloop()
